# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel AI OS Core API
  version: 0.2.5-pilot.8
  description: >-
    PRODUCT-V1.4 / TECH-V0.2-PILOT.8 contract. Production requests use a
    trusted JWT/SSO bearer token; roles, permissions, organization scope and
    active position assignments are resolved by the server. X-Tenant-Id and
    X-Actor-Id are accepted only when DEV_HEADER_AUTH_ENABLED=true for local
    integration. Client-supplied role or organization-scope headers are never
    trusted.
servers:
  - url: http://localhost:8080
security:
  - BearerAuth: []
  - TenantHeader: []
    ActorHeader: []
...
paths:
...
  /api/v1/tasks/targets:
    get:
      tags: [Task Center]
      operationId: listTaskTargets
      summary: List active assignments allowed by the caller's task-specific target policy
      responses:
        '200': { $ref: '#/components/responses/ListResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/tasks/{taskId}:
    get:
      tags: [Task Center]
      operationId: getTask
      summary: Get task details and exact assignment participants
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { $ref: '#/components/responses/ObjectResponse' }
...
  /api/v1/executive-tasks:
    get:
      tags: [Group Executive Task]
      operationId: listExecutiveTasks
      summary: List the chairman's field-minimized group task projection
      description: Requires an active GROUP_CHAIRMAN business assignment, executive-task.read and the tenant feature gate. The list never returns descriptions, result text, reminders or evidence metadata.
      parameters:
        - $ref: '#/components/parameters/AssignmentHeader'
      responses:
        '200':
          description: Field-minimized group task list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ExecutiveTaskSummary' } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Group Executive Task]
      operationId: createExecutiveTask
      summary: Atomically assign a task to a group general manager or vice president
      description: Idempotency-Key must be the canonical UUID text in clientCommandId. Reviewer, source and organization fields are server-derived and rejected if supplied.
      parameters:
        - $ref: '#/components/parameters/AssignmentHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateExecutiveTask' }
      responses:
        '201':
          description: Task created and dispatched to PENDING_ACK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutiveTaskDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/executive-tasks/targets:
    get:
      tags: [Group Executive Task]
      operationId: listExecutiveTaskTargets
      summary: List active group-root GM and vice-president assignments eligible for chairman assignment
      parameters:
        - $ref: '#/components/parameters/AssignmentHeader'
      responses:
        '200':
          description: Eligible target assignments
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ExecutiveTaskTarget' } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/executive-tasks/{taskId}:
    get:
      tags: [Group Executive Task]
      operationId: getExecutiveTask
      summary: Get a field-minimized task detail
      description: Description, result summary, evidence count and reminder times are present only for the current chairman assignment's authoritative directive. Evidence metadata and content are never returned.
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/AssignmentHeader'
      responses:
        '200':
          description: Field-minimized task detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutiveTaskDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
  /api/v1/executive-tasks/{taskId}/actions/approve:
    post:
      tags: [Group Executive Task]
      operationId: approveExecutiveTask
      summary: Approve the current chairman assignment's own authoritative directive
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/AssignmentHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApproveExecutiveTask' }
      responses:
        '200':
          description: Approved task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutiveTaskDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/executive-tasks/{taskId}/actions/rework:
    post:
      tags: [Group Executive Task]
      operationId: reworkExecutiveTask
      summary: Return the current chairman assignment's own authoritative directive for rework
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/AssignmentHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReworkExecutiveTask' }
      responses:
        '200':
          description: Task returned for rework
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutiveTaskDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Trusted production JWT or SSO access token.
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
    ActorHeader:
      type: apiKey
      in: header
      name: X-Actor-Id
      description: Local integration only; requires DEV_HEADER_AUTH_ENABLED=true.
  parameters:
...
    TaskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key; reuse returns the original semantic result.
      schema: { type: string, minLength: 1, maxLength: 200 }
    AssignmentHeader:
      name: X-Assignment-Id
      in: header
      required: false
      description: >-
        Selects the current responsibility-bearing business actor assignment without
        replacing or narrowing account-level roles and permissions.
      schema: { type: string, format: uuid }
  responses:
    ListResponse:
      description: List response
      content:
        application/json:
          schema:
            type: array
            items: { type: object, additionalProperties: true }
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    BadRequest:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Permission denied or organization scope violation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Optimistic version or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [title, status, detail, code]
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum: [AUTHENTICATION_FAILED, BUSINESS_ASSIGNMENT_REQUIRED, BUSINESS_ASSIGNMENT_INVALID_FORMAT, BUSINESS_ASSIGNMENT_FORBIDDEN, BUSINESS_ACTOR_MISMATCH, BUSINESS_ASSIGNMENT_ROLE_MISMATCH, ACCESS_DENIED, REQUEST_INVALID, RESOURCE_NOT_FOUND, FEATURE_NOT_ENABLED, EXECUTIVE_TASK_TARGET_FORBIDDEN, EXECUTIVE_TASK_ACTION_FORBIDDEN, TASK_VERSION_CONFLICT, TASK_STATE_CONFLICT, IDEMPOTENCY_STATE_UNAVAILABLE, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, IDEMPOTENCY_STATE_CONFLICT]
...
    ExecutiveTaskTarget:
      type: object
      additionalProperties: false
      required: [assignmentId, employeeName, positionCode, positionName, organizationId, organizationName]
      properties:
        assignmentId: { type: string, format: uuid }
        employeeName: { type: string }
        positionCode: { type: string, enum: [GROUP_GENERAL_MANAGER, GROUP_VICE_PRESIDENT] }
        positionName: { type: string }
        organizationId: { type: string, format: uuid }
        organizationName: { type: string }
    ExecutiveTaskSummary:
      type: object
      additionalProperties: false
      required: [id, taskNo, title, lifecycleStatus, slaStatus, priority, dueAt, rowVersion, creationSource, orgUnitId, orgUnitName, assigneeAssignmentId, assigneeName, assigneePositionName, reviewerAssignmentId, reviewerName, progress]
      properties:
        id: { type: string, format: uuid }
        taskNo: { type: string }
        title: { type: string }
        lifecycleStatus: { type: string }
        slaStatus: { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
        dueAt: { type: [string, 'null'], format: date-time }
        rowVersion: { type: integer, format: int64, minimum: 0 }
        creationSource: { type: string }
        orgUnitId: { type: string, format: uuid }
        orgUnitName: { type: string }
        assigneeAssignmentId: { type: string, format: uuid }
        assigneeName: { type: string }
        assigneePositionName: { type: string }
        reviewerAssignmentId: { type: string, format: uuid }
        reviewerName: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
    ExecutiveTaskDetail:
      allOf:
        - $ref: '#/components/schemas/ExecutiveTaskSummary'
        - type: object
          properties:
            description: { type: string }
            resultSummary: { type: string }
            evidenceCount: { type: integer, minimum: 0 }
            reminderTimes: { type: array, maxItems: 8, items: { type: string, format: date-time } }
      unevaluatedProperties: false
    CreateExecutiveTask:
      type: object
      additionalProperties: false
      required: [clientCommandId, targetAssignmentId, title, description, dueAt]
      properties:
        clientCommandId: { type: string, format: uuid }
        targetAssignmentId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 240 }
        description: { type: string, minLength: 1 }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT], default: NORMAL }
        dueAt: { type: string, format: date-time }
        reminderTimes: { type: array, maxItems: 8, uniqueItems: true, items: { type: string, format: date-time } }
    ApproveExecutiveTask:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        comment: { type: string }
    ReworkExecutiveTask:
      type: object
      additionalProperties: false
      required: [expectedVersion, reason]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        reason: { type: string, minLength: 1 }
...